refactor: autofit 로직 별도 훅으로 분리
